+++ duckscript_sdk/src/sdk/std/collections/map_keys/mod.rs
@@ -1,0 +1,81 @@
+use crate::utils::pckg;
+use crate::utils::state::{get_handles_sub_state, put_handle};
+use duckscript::types::command::{Command, CommandResult, Commands};
+use duckscript::types::instruction::Instruction;
+use duckscript::types::runtime::StateValue;
+use std::collections::HashMap;
+use std::str;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "MapKeys")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["map_keys".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn requires_context(&self) -> bool {
+        true
+    }
+
+    fn run_with_context(
+        &self,
+        arguments: Vec<String>,
+        state: &mut HashMap<String, StateValue>,
+        _variables: &mut HashMap<String, String>,
+        _output_variable: Option<String>,
+        _instructions: &Vec<Instruction>,
+        _commands: &mut Commands,
+        _line: usize,
+    ) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Map handle not provided.".to_string())
+        } else {
+            let handles_state = get_handles_sub_state(state);
+
+            match handles_state.get(&arguments[0]) {
+                Some(state_value) => match state_value {
+                    StateValue::SubState(ref map) => {
+                        let mut array = vec![];
+
+                        for map_key in map.keys() {
+                            array.push(StateValue::String(map_key.to_string()));
+                        }
+
+                        let key = put_handle(state, StateValue::List(array));
+
+                        CommandResult::Continue(Some(key))
+                    }
+                    _ => CommandResult::Error("Invalid handle provided.".to_string()),
+                },
+                None => CommandResult::Error(
+                    format!("Map for handle: {} not found.", &arguments[0]).to_string(),
+                ),
+            }
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}