+++ duckscript_sdk/src/sdk/std/fs/temp_dir/mod.rs
@@ -1,7 +1,7 @@
 use crate::utils::pckg;
 use duckscript::types::command::{Command, CommandResult};
 use fsio::path::from_path::FromPath;
-use home;
+use std::env;
 
 #[cfg(test)]
 #[path = "./mod_test.rs"]
@@ -14,11 +14,11 @@ pub(crate) struct CommandImpl {
 
 impl Command for CommandImpl {
     fn name(&self) -> String {
-        pckg::concat(&self.package, "GetHomeDirectory")
+        pckg::concat(&self.package, "TempDirectory")
     }
 
     fn aliases(&self) -> Vec<String> {
-        vec!["get_home_dir".to_string()]
+        vec!["temp_dir".to_string()]
     }
 
     fn help(&self) -> String {
@@ -30,13 +30,10 @@ impl Command for CommandImpl {
     }
 
     fn run(&self, _arguments: Vec<String>) -> CommandResult {
-        match home::home_dir() {
-            Some(directory) => {
-                let directory_str = FromPath::from_path(&directory);
-                CommandResult::Continue(Some(directory_str))
-            }
-            None => CommandResult::Error("Unable to find user home directory.".to_string()),
-        }
+        let directory_path = env::temp_dir();
+        let directory = FromPath::from_path(&directory_path);
+
+        CommandResult::Continue(Some(directory))
     }
 }