+++ duckscript_sdk/src/sdk/std/fs/temp_dir/mod_test.rs
@@ -1,0 +1,17 @@
+use super::*;
+use crate::test;
+use crate::test::CommandValidation;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_simple() {
+    test::run_script_and_validate(
+        vec![create("")],
+        "out = temp_dir",
+        CommandValidation::Ignore,
+    );
+}